Use SEXPREC abstract type and subtypes
Each SEXPTYPE has its own Julia type now.  See `subtypes(SEXPREC)` for a list.
In addition to `sexp` and `reval` an `rcopy` generic is exported and `vec` methods
for SEXPRECs are defined.

`rcopy` copies the value of an R object as a Julia object.  There are methods for
`Symbol` and for `ByteString` that evaluate the object in R then copy the value into
a Julia object.

`DataFrame` methods apply to an R data frame (or a `Symbol` or a `ByteString`)

`DataArray` methods apply to R atomic vectors (RealSxp, IntSxp, CplxSxp, LglSxp)

`rprint` invokes the R printing mechanism on an SEXPREC.