Major refactor to use RObject type.
R objects are now represented in 2 ways:
* behind-the-scenes as Ptr{S<:SEXPREC}
* user-facing RObject{S<:SEXPREC}, which provide GC-protected pointers

The advantage of this is that it is now clear whether or not an R object is protected from garbage collection.

In order to convert a Julia object to an R object:
```
RObject(x) # GC-protected
sexp(x)    # returns Ptr, not protected
```
as a result, `sexp` is no longer exported. To do the reverse, use `rcopy`, with optional type:
```
rcopy(x) # default conversion
rcopy(T,x) # specific conversion
```
This also avoids the need to overload `DataFrames(::Symbol)` and such.

Other changes:
* SEXPRECs are now immutable
* whenever possible, the correct SEXPREC type is returned from functions, to improve type information
* nice wrappers to R functions (CAR, CDR, allocArray, setAttrib!, etc.)
* correctly call `Rf_xlength` for vectors.
* Use of `Base.vec` has been replaced by `unsafe_vec` (not exported)
* Use of `Base.names` has been replaced by `getNames` (exported)
* `protect` and `unprotect` are used to ensure Ptr{S<:SEXPREC} are not GC-ed.