Enable precompilation, bump Julia version requirement to 0.4
This modifies the internals to allow for package precompilation. Due to some of the necessary changes (e.g. we can no longer use `const` pointers), this is a breaking change, though hopefully the "high-level" interface should still work.

This also changes the library search functionality so that we no longer need to use `Pkg.build`.