Allow for dispatching on the R functions
In the current setup, RCall returns a closure in Julia, which means it's hard to know in Julia if something is a function from R. This makes it really hard to workaround things that are weird in R, like is seen in https://github.com/SciML/diffeqr/issues/39. This changes it to use a callable struct instead of a closure in order to allow for dispatching.