Fix REPL issue on Julia nightly
The `hint` keyword was added to the API of `complete_line`, and a
corresponding positional argument was added to `bslash_completions`.
This change adds the `hint` keyword to our method of `complete_line` and
passes it along to `bslash_completions`, which is defined locally to
accept and ignore the `hint` argument for earlier Julia versions.